This file describes what have been added to ALARA
10/18/2007

Exposure dose capability is added to ALARA. This functionality has not been validated. The user can invoke this function by using a keyword "exposure" in the output block. Following the keyword are gamma library name, gamma attenuation file, a radius of a cylinder and a distance from a cylinder. ALARA will use inifinite line approximation to calculate the exposure dose. The unit is mR/hr.

output
   exposure fendl2bin ANS6_4_3 1 100
end


10/19/2007

The output block for the exposure dose is removed. Now, the user can invoke the calculation by using a keyword "dose" (the same as contact dose).

output
   dose contact fenl2bin ANS6_4_3        # calculate contact dose using a semi-inifinite medium approximation
   dose line fendl2bin ANS6_4_3 1 100    # calculate exposure dose using an infinite line approximation. This will give the same result as the one on 10/18/07
end